Information error
sir this should be adding two integer numbers, if not it should be a float or a double. here I used double. and its better to use adding two consequent numbers

+++ Add Two Numbers FLOW001.c
@@ -2,10 +2,10 @@
 
 main()
 {
-    int t; scanf("%d",&t);
+    double t; scanf("%f",&t);
     while(t--)
     {
     	long int a,b;
-    	scanf("%ld %ld",&a,&b); printf("%ld\n",a+b);
+    	scanf("%f %f",&a,&b); printf("%f\n",a+b);
     }
 }
